requirements.txt
Flask==0.10.1
Flask-DebugToolbar==0.8.0
Flask-Mail==0.9.0
Flask-SQLAlchemy==1.0
Flask-Script==0.6.2
Flask-WTF==0.9.1
GeoAlchemy==0.7.2
Jinja2==2.7.1
MarkupSafe==0.18
Unidecode==0.04.14
WTForms==1.0.4
Werkzeug==0.9.4
amqp==1.3.3
anyjson==0.3.3
argparse==1.2.1
git+https://github.com/behave/behave.git
billiard==3.3.0.13
blinker==1.3
celery==3.1.7
foursquare==2013.12.10
geojson==1.0.1
httplib2==0.8
itsdangerous==0.23
kombu==3.0.8
oauth2==1.5.211
oauthlib==0.6.0
parse==1.6.3
pyshp==1.2.0
python-dateutil==2.2
python-openid==2.2.5
python-slugify==0.0.6
pytz==2013.8
redis==2.9.0
requests==2.1.0
requests-oauthlib==0.4.0
simplejson==3.3.0
SQLAlchemy==0.9.1
six==1.4.1
tweepy==2.1
wsgiref==0.1.2